==> scripts/4.1-MonthlyWeatherCleanDataEco.Rmd <==
---
title: "Monthly weather by Eco"
author: "Anna Talucci"
date: "6/14/2021"
output: html_document
---


# Clear workspace

```{r}
rm(list=ls()) 
```


# Overview
[Terra Climate](https://developers.google.com/earth-engine/datasets/catalog/IDAHO_EPSCOR_TERRACLIMATE#bands)

Name    |	Units   |	Min	  |  Max   |	Scale	    | Description
def	    |   mm	  |   0*  |  4548* |	0.1	      | Climate water deficit, derived using a one-dimensional soil water balance model
pdsi    |		       | -4317*	| 3418*	| 0.01	     |Palmer Drought Severity Index

pr      |	mm	    |0*	      |7245*	|	         |Precipitation accumulation
tmmx	 |  °C	    |-670*	  | 576*	| 0.1	     |Maximum temperature

# Packages


```{r include=FALSE}

library(dplyr)
library(tidyr)
library(ggplot2)
library(cowplot)

```



# By Ecozone

```{r}
mw = read.csv("../data/MonthlyWeatherEco/2020_meanMonthEco.csv") 
```




```{r}
head(mw)

```





```{r}
data4 = mw %>% rename(def_01=X202001_def, 
           
                pr_01=X202001_pr, 
                tmmx_01=X202001_tmmx, 
                def_02=X202002_def, 
           
                pr_02=X202002_pr, 
                tmmx_02=X202002_tmmx, 
                def_03=X202003_def, 
            
                pr_03=X202003_pr, 
                tmmx_03=X202003_tmmx,
                def_04=X202004_def, 
          
                pr_04=X202004_pr, 
                tmmx_04=X202004_tmmx,
                def_05=X202005_def, 
       
                pr_05=X202005_pr, 
                tmmx_05=X202005_tmmx,
                def_06=X202006_def, 
         
                pr_06=X202006_pr, 
                tmmx_06=X202006_tmmx,
                def_07=X202007_def, 
             
                pr_07=X202007_pr, 
                tmmx_07=X202007_tmmx,
                def_08=X202008_def, 
               
                pr_08=X202008_pr, 
                tmmx_08=X202008_tmmx,
                def_09=X202009_def, 
               
                pr_09=X202009_pr, 
                tmmx_09=X202009_tmmx,
                def_10=X202010_def, 
                 
                pr_10=X202010_pr, 
                tmmx_10=X202010_tmmx,
                def_11=X202011_def, 
                 
                pr_11=X202011_pr, 
                tmmx_11=X202011_tmmx,
                
                def_12=X202012_def,
                pr_12= X202012_pr,
                tmmx_12 = X202012_tmmx) 

data4
```

```{r}
data5 = data4 %>%
 pivot_longer(
   cols = c("def_01", "pr_01", "tmmx_01", "def_02",  "pr_02" ,"tmmx_02", "def_03",  "pr_03", "tmmx_03","def_04",  "pr_04", "tmmx_04", "def_05", "pr_05", "tmmx_05", "def_06", "pr_06",  "tmmx_06", "def_07", "pr_07", "tmmx_07", "def_08", "pr_08",  "tmmx_08", "def_09", "pr_09", "tmmx_09","def_10", "pr_10", "tmmx_10", "def_11",  "pr_11","tmmx_11", "def_12",  "pr_12","tmmx_12")  ,
   names_to = "wm",
   values_to = "measures") %>%
  dplyr::select(ECO_NAME, eco_code, OBJECTID, wm, measures) %>%
  separate(wm, c("weather", "month")) %>%
  pivot_wider(names_from = weather, values_from = measures)

data5
```

```{r}
data6 = data5 %>%
  mutate(def1=round(def*0.1,3)) %>%
  mutate(tmmx1=round(tmmx*0.1,3)) %>%
  dplyr::select(ECO_NAME, eco_code, OBJECTID, month, pr, def1, tmmx1) %>%
  rename(def=def1, tmmx=tmmx1) %>%
  mutate(FireYr = "2020")

data6
  
   
```


```{r}
write.csv(data6, '../outputs/CleanWeatherEco/2020_MeanMonth.csv',  row.names = FALSE)
```


**THE END**
